Use *ST suffix for system tests
