+++ java/system-test/system-test-suite/src/test/java/sleeper/systemtest/suite/GarbageCollectionST.java
@@ -42,7 +42,7 @@ import static sleeper.core.testutils.printers.FileReferencePrinter.printFiles;
 import static sleeper.systemtest.suite.fixtures.SystemTestInstance.MAIN;
 
 @SystemTest
-public class GarbageCollectionIT {
+public class GarbageCollectionST {
     @TempDir
     private Path tempDir;
