+++ java/system-test/system-test-suite/src/test/java/sleeper/systemtest/suite/PythonBulkImportST.java
@@ -19,45 +19,46 @@ package sleeper.systemtest.suite;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import sleeper.core.record.Record;
-import sleeper.core.util.PollWithRetries;
 import sleeper.systemtest.dsl.SleeperSystemTest;
 import sleeper.systemtest.dsl.extension.AfterTestPurgeQueues;
+import sleeper.systemtest.dsl.extension.AfterTestReports;
+import sleeper.systemtest.dsl.reporting.SystemTestReports;
 import sleeper.systemtest.suite.testutil.SystemTest;
 
-import java.time.Duration;
 import java.util.Map;
+import java.util.stream.LongStream;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static sleeper.configuration.properties.instance.CdkDefinedInstanceProperty.BULK_IMPORT_EMR_JOB_QUEUE_URL;
+import static sleeper.configuration.properties.instance.CdkDefinedInstanceProperty.BULK_IMPORT_EMR_SERVERLESS_JOB_QUEUE_URL;
 import static sleeper.configuration.properties.table.TableProperty.BULK_IMPORT_MIN_LEAF_PARTITION_COUNT;
 import static sleeper.systemtest.suite.fixtures.SystemTestInstance.MAIN;
 
 @SystemTest
-public class EmrBulkImportIT {
+public class PythonBulkImportST {
 
     @BeforeEach
-    void setUp(SleeperSystemTest sleeper, AfterTestPurgeQueues purgeQueues) {
+    void setup(SleeperSystemTest sleeper, AfterTestReports reporting, AfterTestPurgeQueues purgeQueues) {
+        reporting.reportIfTestFailed(SystemTestReports.SystemTestBuilder::ingestTasksAndJobs);
+        purgeQueues.purgeIfTestFailed(BULK_IMPORT_EMR_SERVERLESS_JOB_QUEUE_URL);
         sleeper.connectToInstance(MAIN);
-        purgeQueues.purgeIfTestFailed(BULK_IMPORT_EMR_JOB_QUEUE_URL);
     }
 
     @Test
-    void shouldBulkImportOneRecordWithEmrByQueue(SleeperSystemTest sleeper) {
+    void shouldBulkImportFilesFromS3(SleeperSystemTest sleeper) {
         // Given
         sleeper.updateTableProperties(Map.of(BULK_IMPORT_MIN_LEAF_PARTITION_COUNT, "1"));
-        Record record = new Record(Map.of(
-                "key", "some-id",
-                "timestamp", 1234L,
-                "value", "Some value"));
+        sleeper.sourceFiles()
+                .createWithNumberedRecords("file1.parquet", LongStream.range(0, 100))
+                .createWithNumberedRecords("file2.parquet", LongStream.range(100, 200));
 
         // When
-        sleeper.sourceFiles().create("file.parquet", record);
-        sleeper.ingest().bulkImportByQueue().sendSourceFiles(BULK_IMPORT_EMR_JOB_QUEUE_URL, "file.parquet")
-                .waitForJobs(PollWithRetries.intervalAndPollingTimeout(Duration.ofSeconds(30), Duration.ofMinutes(30)));
+        sleeper.pythonApi()
+                .bulkImport().fromS3("file1.parquet", "file2.parquet")
+                .waitForJobs();
 
         // Then
         assertThat(sleeper.directQuery().allRecordsInTable())
-                .containsExactly(record);
+                .containsExactlyElementsOf(sleeper.generateNumberedRecords(LongStream.range(0, 200)));
+        assertThat(sleeper.tableFiles().references()).hasSize(1);
     }
 }
