+++ java/system-test/system-test-suite/src/test/java/sleeper/systemtest/suite/PythonIngestST.java
@@ -18,6 +18,7 @@ package sleeper.systemtest.suite;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.io.TempDir;
 
 import sleeper.systemtest.dsl.SleeperSystemTest;
 import sleeper.systemtest.dsl.extension.AfterTestPurgeQueues;
@@ -25,36 +26,71 @@ import sleeper.systemtest.dsl.extension.AfterTestReports;
 import sleeper.systemtest.dsl.reporting.SystemTestReports;
 import sleeper.systemtest.suite.testutil.SystemTest;
 
-import java.util.Map;
+import java.nio.file.Path;
 import java.util.stream.LongStream;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static sleeper.configuration.properties.instance.CdkDefinedInstanceProperty.BULK_IMPORT_EMR_SERVERLESS_JOB_QUEUE_URL;
-import static sleeper.configuration.properties.table.TableProperty.BULK_IMPORT_MIN_LEAF_PARTITION_COUNT;
+import static sleeper.configuration.properties.instance.CdkDefinedInstanceProperty.INGEST_JOB_QUEUE_URL;
 import static sleeper.systemtest.suite.fixtures.SystemTestInstance.MAIN;
 
 @SystemTest
-public class PythonBulkImportIT {
+public class PythonIngestST {
+    @TempDir
+    private Path tempDir;
 
     @BeforeEach
     void setup(SleeperSystemTest sleeper, AfterTestReports reporting, AfterTestPurgeQueues purgeQueues) {
         reporting.reportIfTestFailed(SystemTestReports.SystemTestBuilder::ingestTasksAndJobs);
-        purgeQueues.purgeIfTestFailed(BULK_IMPORT_EMR_SERVERLESS_JOB_QUEUE_URL);
+        purgeQueues.purgeIfTestFailed(INGEST_JOB_QUEUE_URL);
         sleeper.connectToInstance(MAIN);
     }
 
     @Test
-    void shouldBulkImportFilesFromS3(SleeperSystemTest sleeper) {
+    void shouldBatchWriteOneFile(SleeperSystemTest sleeper) {
         // Given
-        sleeper.updateTableProperties(Map.of(BULK_IMPORT_MIN_LEAF_PARTITION_COUNT, "1"));
+        sleeper.localFiles(tempDir)
+                .createWithNumberedRecords("file.parquet", LongStream.range(0, 100));
+
+        // When
+        sleeper.pythonApi()
+                .ingestByQueue().uploadingLocalFile(tempDir, "file.parquet")
+                .invokeTask().waitForJobs();
+
+        // Then
+        assertThat(sleeper.directQuery().allRecordsInTable())
+                .containsExactlyElementsOf(sleeper.generateNumberedRecords(LongStream.range(0, 100)));
+        assertThat(sleeper.tableFiles().references()).hasSize(1);
+    }
+
+    @Test
+    void shouldIngestTwoFilesFromS3(SleeperSystemTest sleeper) {
+        // Given
         sleeper.sourceFiles()
                 .createWithNumberedRecords("file1.parquet", LongStream.range(0, 100))
                 .createWithNumberedRecords("file2.parquet", LongStream.range(100, 200));
 
         // When
         sleeper.pythonApi()
-                .bulkImport().fromS3("file1.parquet", "file2.parquet")
-                .waitForJobs();
+                .ingestByQueue().fromS3("file1.parquet", "file2.parquet")
+                .invokeTask().waitForJobs();
+
+        // Then
+        assertThat(sleeper.directQuery().allRecordsInTable())
+                .containsExactlyElementsOf(sleeper.generateNumberedRecords(LongStream.range(0, 200)));
+        assertThat(sleeper.tableFiles().references()).hasSize(1);
+    }
+
+    @Test
+    void shouldIngestDirectoryFromS3(SleeperSystemTest sleeper) {
+        // Given
+        sleeper.sourceFiles()
+                .createWithNumberedRecords("test-dir/file1.parquet", LongStream.range(0, 100))
+                .createWithNumberedRecords("test-dir/file2.parquet", LongStream.range(100, 200));
+
+        // When
+        sleeper.pythonApi()
+                .ingestByQueue().fromS3("test-dir")
+                .invokeTask().waitForJobs();
 
         // Then
         assertThat(sleeper.directQuery().allRecordsInTable())
