+++ java/system-test/system-test-suite/src/test/java/sleeper/systemtest/suite/TableMetricsST.java
@@ -39,7 +39,7 @@ import static sleeper.systemtest.suite.fixtures.SystemTestInstance.MAIN;
 import static sleeper.systemtest.suite.fixtures.SystemTestSchema.DEFAULT_SCHEMA;
 
 @SystemTest
-public class TableMetricsIT {
+public class TableMetricsST {
 
     @TempDir
     private Path tempDir;
