+++ java/system-test/system-test-suite/src/test/java/sleeper/systemtest/suite/approvals/CompactionOnEC2ST.MergeFiles.shouldCompactFilesUsingBasicCompactionStrategy.approved.txt
