+++ java/system-test/system-test-suite/src/test/java/sleeper/systemtest/suite/approvals/CompactionOnFargateST.MergeFiles.shouldCompactFilesUsingBasicCompactionStrategy.approved.txt
