+++ java/system-test/system-test-suite/src/test/java/sleeper/systemtest/suite/approvals/GarbageCollectionST.shouldGarbageCollectFilesAfterCompaction.approved.txt
